Split Settings display and tweaks panels
