+++ service-worker.js
@@ -252,6 +252,8 @@ const APP_SHELL = [
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
   '/js/settings.js',
+  '/js/settings-display-panel.js',
+  '/js/settings-event-target.js',
   '/js/settings-loader.js',
   '/js/settings-privacy.js',
   '/js/settings-data.js',
@@ -261,6 +263,7 @@ const APP_SHELL = [
   '/js/settings-runtime-bridge.js',
   '/js/settings-sync-panel.js',
   '/js/settings-sync-panel-impl.js',
+  '/js/settings-tweaks.js',
   '/js/settings-agent-access-panel.js',
   '/js/feedback.js',
   '/js/tour-runtime.js',
