+++ tsconfig.checkjs.json
@@ -237,10 +237,13 @@
     "js/recommendations-runtime.js",
     "js/recommendations.js",
     "js/settings.js",
+    "js/settings-display-panel.js",
+    "js/settings-event-target.js",
     "js/settings-loader.js",
     "js/settings-privacy.js",
     "js/settings-runtime.js",
     "js/settings-runtime-bridge.js",
+    "js/settings-tweaks.js",
     "js/service-worker-update.js",
     "js/shell-actions.js",
     "js/startup-foundation.js",
