+++ tsconfig.json
@@ -251,10 +251,13 @@
     "js/schema.js",
     "js/schema-environment.js",
     "js/settings.js",
+    "js/settings-display-panel.js",
+    "js/settings-event-target.js",
     "js/settings-privacy.js",
     "js/settings-runtime.js",
     "js/settings-sync-panel.js",
     "js/settings-sync-panel-impl.js",
+    "js/settings-tweaks.js",
     "js/settings-agent-access-panel.js",
     "js/shell-actions.js",
     "js/silhouette-paths.js",
